// hr-frontend/src/lib/api/generated/model/contractTypeSetting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ContractTypeSetting {
  contractTypeId?: number;
  code?: string;
  name?: string;
  description?: string;
  active?: boolean;
  usageCount?: number;
  createdAt?: string;
  updatedAt?: string;
}